% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbitax.R
\docType{data}
\name{cacheEnv}
\alias{cacheEnv}
\title{Environment keeping cached NCBI tree}
\format{An object of class \code{environment} of length 0.}
\usage{
cacheEnv
}
\description{
Environment keeping cached NCBI tree
}
\keyword{internal}
