% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbitax.R
\name{download.ncbitax}
\alias{download.ncbitax}
\title{Download NCBI Taxonomy Files}
\usage{
download.ncbitax(destdir,
  baseurl = "ftp://ftp.ncbi.nlm.nih.gov/pub/taxonomy/",
  fname = "taxdump.tar.gz")
}
\arguments{
\item{destdir}{Directory to unpack files into. A temporary directory will be
created if this is not provided.}

\item{baseurl}{URL from which taxdump should be downloaded}

\item{fname}{Filename to expect at \code{baseurl}}
}
\value{
Directory name
}
\description{
Downloads the NCBI Taxonomy dump and unpacks it into a temporary directory.
}
\examples{
dirname <- download.ncbitax()

}
\seealso{
\code{\link{read.ncbi}}
}
