% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbitax.R
\name{ncbi.group}
\alias{ncbi.group}
\title{Yield summary group name for taxids}
\usage{
ncbi.group(taxid, ranks = c(), nodes = c(), ncbi = read.ncbi())
}
\arguments{
\item{taxid}{Set of NCBI taxonomy IDs}

\item{ranks}{Vector of rank names}

\item{nodes}{Vector of clade names}

\item{ncbi}{Explicitly provide NCBI taxonomy as generated by
\code{\link{read.ncbi}}.}
}
\value{
Vector of clade names
}
\description{
Computes summary group names for a set of tax ids such that the group names
are either in the set provided by \code{nodes} or at one of the ranks provided
by \code{ranks}.
}
\examples{

ncbi.group(c(9606, 22, 9), ranks=c("superkingdom"),
           nodes=c("Homo sapiens", "synthetic construct"))

}
