% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbitax.R
\name{ncbi.lineage}
\alias{ncbi.lineage}
\title{Computes lineage from NCBI tax_id}
\usage{
ncbi.lineage(tax_id, ncbi = read.ncbi())
}
\arguments{
\item{tax_id}{NCBI taxid(s). Passing a vector is highly recommended over
calling \code{ncbi.lineage} multiple times.}

\item{ncbi}{Explicitly provide NCBI taxonomy as generated by
\code{\link{read.ncbi}}.}
}
\value{
List of NCBI taxids. If \code{tax_id} is a vector, a list of lists
  will be returned with all lists filled to matching lengths using the id of
  the root node (1).
}
\description{
Computes lineage from NCBI tax_id
}
\examples{

ncbi.lineage(9606)

}
\seealso{
\code{\link{ncbi.path}}
}
