% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbitax.R
\name{ncbi.norm_path}
\alias{ncbi.norm_path}
\title{Computes normalized taxonomy path from NCBI tax_id}
\usage{
ncbi.norm_path(t_tax_id, ncbi = read.ncbi(), ranks = factor(c("species",
  "genus", "family", "order", "class", "phylum", "superkingdom")))
}
\description{
Computes normalized taxonomy path from NCBI tax_id
}
