% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbitax.R
\name{ncbi.path}
\alias{ncbi.path}
\title{Computes taxonomy path from NCBI tax_id}
\usage{
ncbi.path(tax_id, ranks, ncbi = read.ncbi())
}
\arguments{
\item{tax_id}{NCBI taxid(s). Passing a vector is highly recommended over
calling \code{ncbi.path} multiple times.}

\item{ranks}{Filter resulting vector to only include the listed ranks.
Passing \code{"standard"} will filter the lineages to contain only the
levels species, genus, family, order, class, phylum and superkingdom
(domain).}

\item{ncbi}{Explicitly provide NCBI taxonomy as generated by
\code{\link{read.ncbi}}.}
}
\value{
List of NCBI taxonomy node names. If \code{tax_id} is a vector, a
  list of lists will be returned.
}
\description{
Computes taxonomy path from NCBI tax_id
}
\examples{
ncbi.path(9606)
ncbi.path(9606, ranks = "standard")
ncbi.path(9606, ranks = c("class", "phylum"))

}
\seealso{
\code{\link{ncbi.lineage}}
}
