% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbitax.R
\name{ncbi.path}
\alias{ncbi.path}
\title{Computes taxonomy path from NCBI tax_id}
\usage{
ncbi.path(t_tax_id, ncbi = read.ncbi())
}
\arguments{
\item{ncbi}{Explicitly provide NCBI taxonomy as generated by
\code{\link{read.ncbi}}.}

\item{tax_id}{NCBI taxid(s). Passing a vector is highly recommended over
calling \code{ncbi.path} multiple times.}
}
\value{
List of NCBI taxonomy node names. If \code{tax_id} is a vector, a
  list of lists will be returned.
}
\description{
Computes taxonomy path from NCBI tax_id
}
\examples{
ncbi.path(9606)
}
\seealso{
\code{\link{ncbi.lineage}}
}
